* Code refactoring
