+++ src/main/java/ru/alemakave/mfstock/MFStockApplication.java
@@ -4,7 +4,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.builder.SpringApplicationBuilder;
-import ru.alemakave.mfstock.configs.service.MFStockConfigLoader;
+import ru.alemakave.mfstock.configs.MFStockConfigLoader;
 
 import javax.annotation.PostConstruct;
