+++ src/main/java/ru/alemakave/mfstock/model/configs/DBConfigs.java
@@ -1,4 +1,4 @@
-package ru.alemakave.mfstock.configs.model;
+package ru.alemakave.mfstock.model.configs;
 
 import java.util.Arrays;
