+++ src/main/java/ru/alemakave/mfstock/service/DBServiceImpl.java
@@ -7,18 +7,20 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.ConfigurableApplicationContext;
 import org.springframework.stereotype.Service;
-import ru.alemakave.mfstock.configs.model.DBConfigsColumns;
-import ru.alemakave.mfstock.configs.service.MFStockConfigLoader;
+import ru.alemakave.mfstock.configs.MFStockConfigLoader;
 import ru.alemakave.mfstock.exceptions.DBException;
+import ru.alemakave.mfstock.model.configs.DBConfigsColumns;
 import ru.alemakave.mfstock.model.table.Table;
 import ru.alemakave.mfstock.model.table.TableCell;
 import ru.alemakave.mfstock.model.table.TableRow;
-import ru.alemakave.mfstock.model.telegram_bot.TelegramCachePhotoFilesManager;
+import ru.alemakave.mfstock.telegram_bot.TelegramCachePhotoFilesManager;
 import ru.alemakave.mfstock.utils.PageUtils;
 import ru.alemakave.mfstock.utils.TableUtils;
 
 import javax.annotation.PostConstruct;
-import java.io.*;
+import java.io.File;
+import java.io.FileInputStream;
+import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
