+++ src/main/java/ru/alemakave/mfstock/service/StickerServiceImpl.java
@@ -6,11 +6,11 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.context.ConfigurableApplicationContext;
 import org.springframework.stereotype.Service;
-import ru.alemakave.mfstock.configs.service.MFStockConfigLoader;
-import ru.alemakave.mfstock.model.generators.CellStickerGenerator;
-import ru.alemakave.mfstock.model.generators.NomPartyStickerGenerator;
-import ru.alemakave.mfstock.model.generators.NomSerStickerGenerator;
-import ru.alemakave.mfstock.model.generators.NomStickerGenerator;
+import ru.alemakave.mfstock.configs.MFStockConfigLoader;
+import ru.alemakave.mfstock.generators.CellStickerGenerator;
+import ru.alemakave.mfstock.generators.NomPartyStickerGenerator;
+import ru.alemakave.mfstock.generators.NomSerStickerGenerator;
+import ru.alemakave.mfstock.generators.NomStickerGenerator;
 import ru.alemakave.mfstock.model.json.CellSticker;
 import ru.alemakave.mfstock.model.json.NomPartySticker;
 import ru.alemakave.mfstock.model.json.NomSerSticker;
