+++ src/main/java/ru/alemakave/mfstock/telegram_bot/actions/GetPhotoTelegramAction.java
@@ -1,7 +1,7 @@
-package ru.alemakave.mfstock.model.telegram_bot.actions;
+package ru.alemakave.mfstock.telegram_bot.actions;
 
 import org.telegram.telegrambots.meta.api.objects.Update;
-import ru.alemakave.mfstock.service.telegram_bot.TelegramBot;
+import ru.alemakave.mfstock.telegram_bot.TelegramBot;
 import ru.alemakave.telegram_bot_utils.actions.TelegramReceiveMessageAction;
 
 import java.io.File;
