+++ src/main/java/ru/alemakave/telegram_bot_utils/actions/TelegramReceiveMessageAction.java
@@ -1,8 +1,8 @@
 package ru.alemakave.telegram_bot_utils.actions;
 
 import org.telegram.telegrambots.meta.api.objects.Update;
-import ru.alemakave.mfstock.model.telegram_bot.TelegramCachePhotoFilesManager;
-import ru.alemakave.mfstock.service.telegram_bot.TelegramBot;
+import ru.alemakave.mfstock.telegram_bot.TelegramBot;
+import ru.alemakave.mfstock.telegram_bot.TelegramCachePhotoFilesManager;
 
 public abstract class TelegramReceiveMessageAction implements ITelegramReceiveAction {
     public final TelegramBot bot;
